Move from Docker Hub to GitHub Container Registry
Depends on prebuild/docker-images#19. Effectively fixes a critical
bug (prebuild/docker-images#17) for dockcross-based images. Those
are:

- `linux-armv6`
- `linux-armv7`
- `linux-arm64`
- `android-armv7`
- `android-arm64`

Also pins image versions (to version 1) by default, which is now
possible because the images are tagged with version numbers in
addition to the `latest` tag.

Image descriptions have moved to the prebuild/docker-images
repository.